wifimonster/requirements.txt
PyQt>=4.8,<5.0
scapy>=2.2.0-dev
